feat(GitOpsCluster): Propagate ManagedCluster labels to created secrets (#51)
This propagates labels set on ManagedClusters to the secrets created
by the GitOpsCluster controller. The labels can e.g. be used in ArgoCD
to select different overlays when using Kustomize.

The reconciliation code of the GitOpsCluster controller is changed so
secrets are always updated, whether they are blank secrets or not. This
allows to propagate changes to the labels of a ManagedCluster after a
secret was already created.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>
(cherry picked from commit 8eaf81c076caf55cd3a9e25184a9daca7cee8695)